--help: just <co-tree> instead of <branch-co-tree>
I remembered that many git users don't even
know that "detached HEAD" is a thing, let
alone what it means or how it can happen,
and that the overwhelmingly typical case
will be that a co-tree is on a branch, not
in a detached HEAD state. So distinguishing
branch co-trees from detached-HEAD co-trees
tersely in the usage text is probably worse
for understanding rather than helpful. If
you actually hit the "not a branch" error
when trying to use `--base` on a co-tree in
detached HEAD state, then it's time to find
out what the distinction is, but until then,
you might never need to worry about it.