Fix --init when whole directories are ignored
The prior easy fix with `git stash --all`
worked in the cases where directories had
some ignored files, but broke when there
are directories with only ignored files
(those directories were properly stashed,
but were unstashed back outside of the new
worktree for some reason).

This is the more complete/thorough fix that
I was trying to avoid spending the time to
get right, but the upside is that it also
avoids the inefficiency of the prior fix
(copying ignored files into Git objects,
deleting them, and copying them back out).