Drop the lock-worktree-of-base-branch concept
This was, if I remember right, to protect the
user from doing a `git branch -d {{base branch}}`
without realizing that will break the ref in the
base/main/root directory.

But honestly, it's just more elegant, simple,
and versatile to not have the lock, to let the
user do `git cotree --delete` on any worktree,
even the last one, and just check to avoid the
branch deletion if it's also the base branch.
Clean.

If this is ever a problem, could look at having
the other co-tree operations handle it.