Full "smart" co-tree name/directory resolution
I have some mixed feelings about the complexity
of this: will it have downsides for programmatic
use? This degree of being contextually-dependent
can often trip up the stuff built on top of it.

But! Ths UX benefit is huge. Names will basically
always do what you mean now. When calling cotree
from outside an existing worktree, the name works
like a relative path from the current location.
When calling cotree from calling cotree from
inside a worktree, a name will be relative to the
repository root, unless it starts with "../":
leading "../" are resolved, and if that takes you
out of the current worktree (which it always will
if the "../" was the result of interactive aim at
an existing worktree or valid location for one)
it will be treated as relative to your location
again.

Of course, the ideal would be to basically resolve
all possible paths (incl. symlinks, ".." parts in
the middle or end rather than the start, etc), and
I suspect in due time it'll be obvious to me how
to do that elegantly (there's probably a way that
takes less code than my current approach..., but a
simple `cd` currently seems like it would have too
much fiddling to do around directories that don't
exist, plus maybe other cases - so I'll leave that
for someone else or a future me with a fresh mind
for the problem).