Improved subdirectory cleanup UX
Make a best-effort attempt to portably and without
error message matching (locales!) detect when an
`rmdir` fails due to the non-issue of other stuff
remaining in the directory, and suppress just that
error message.

Profoundly idiotic and an ethical+intellectual
failing of all responsible that we haven't had
a practically portable `rmdir` flag for that
for the last twenty years or so. How do you care
or think so little about user+dev experience and
error awareness+handling? Spurious non-errors are
noise and at every interface boundary we should
provide a way to unambiguously filter just them
out so that the real errors can be seen/handled.