Use `return` instead of `exit` where possible
Not using `return` was an Extreme Portability
habit which I've forgotten the origin of and
so at this point it might be cargo-culted and
I might've been doing it in ways that don't
even in principle improve portability.

Either way, `return` is perfectly portable in
practice today.

Even if it wasn't, that's not a good reason
to not use it. Even if anyone ever actually
used a shell where it didn't work, a basic
search could replace the returns with exits.

In the meantime, it makes the code clearer,
more maintanable and easy to reason about,
allows getting rid of a subshell, and lets
me reuse the current functions more easily.