Normalize redundant `//` in co-tree name/path
Why do this? `git branch a//b` or `git checkout
-b a//b` would error out with "bad branch name".
Why not just immitate them and force users to
fix their stuff to not have `//` in their paths?
Because cotree lets you treat branches like
directories, and in a sense its whole Thing is
that conceptually it makes a whole lot of sense
for each branch to be a directory: and paths in
the file system can be referred to with `//`
treated as `/` (and this is a good thing since
it allows writing simple code that doesn't have
to do exactly what cotree has to do here).

P.S. the simple logic used here would be wrong
for absolute paths\*, but for now cotree does
not support absolute paths anyway.

\* Technically if a path *starts* with exactly
   two "/" characters (but no more) then that's
   potentially actually meaningful (per POSIX)
   and thus normalization for absolute paths
   needs to work like this to be fully correct:

    `/foo` -> `/foo`
    `//foo` -> `//foo`
    `///foo` -> `/foo`