Allow any commit-ish as base, from co-tree or not
I am actually going to go back to disallowing
non-branches for the base in the next commit,
but I want to save this in the commit history
so that I can go back to it if I change my
mind yet again.

The biggest advantage is that it would allow
doing something like `git cotree --base v1.2.3`
to make new cotree branches start at the last
release commit (if you're doing a workflow like
that instead of the more common "merge early,
merge often" approach where you're branching
from your latest development head). But how
common/likely is that?

The other big advantage is that it's doing the
"define errors out of existence" best practice
to allow a non-branch base... although it's not
entirely clear if that principle beats out the
benefit of making it harder to get in a state
that doesn't work as expected in this case (how
many people feel fully comfortable and fluent
in a detached head state? how many would expect
that when you set the base to a tag, delete and
recreate the tag, and create a new cotree from
the default base, it will branch from the tag's
commit at the time when base was set?)...

P.S. One nicety here would be to show a message
that the base is being set to a commit hash if
that's what's happening.