Allow base branch either by name or from co-tree
    git cotree --base (<branch> | <co-tree>)

Previously (except in the commit right before
this one) the base had to be a branch which was
checked out into a cotree. This was a hold-over
from when the base branch cotree was locked to
help users not accidentally delete it. (Now, if
I was going to add that protection, I would use
a private worktree like `.git-cotree-safety`
with a sparse checkout to hold the branch open,
and probably lock that.) Currently, the locking
and protection isn't being done at all (see
commit 8ab578e90f553c4a6d0693a560c178556004f409)
and I want to let that ride to see if it ever
becomes a real problem.

This reverts the "allow any commit-ish" part of
commit e9b7a098625377211f350ded3f60f3f31d1fc840
(see that commit for reasoning on that, but in
short: the user-experience is so dubious there,
with little non-obvious pitfalls and breakages,
that I'm not sure it's sufficiently valuable to
allow arbitrary commits as bases). Note: also,
if we do allow any commit-ish for base in the
future, we should probably make `--init` work
from detached head state too (I haven't given
much thought to that).