Tiny optimization for unused log output
Just output nothing in the first place
instead of capturing or discarding the
output.

Using `--pretty=format:` instead of
`--pretty=` because the latter is
better-documented in the `git-log`
manpage.

Using a redundant `''` for the humans,
to whom it's not redundant but a useful
cue that the format is an empty string
(no need for their mind to do
shell-specific parsing and inference).

(Using `git log` to get Git's native
user-facing error if we're for example
in a fresh repo with no commits yet.)