Factor logic to find cotree root into function
The cotree root (by which I mean the repo's
root, but specifically when the repo has the
cotree directory layout) is tricky to find
correctly, and the current logic only works
when gitdir is at .git in the repo root. Not
when any of the myriad ways of having it
elsewhere are in effect (gitlink file at .git
for the common dir, environment variables like
GIT_DIR/GIT_COMMON_DIR/GIT_WORK_TREE and so on.
Pulling it out into its own function because
it's about to get more complex to fix it.