Add collapsing from cotrees back to normal
    git cotree --collapse <co-tree>

The last major feature necessary for a great
user experience: when you're done with using
multiple worktrees for a while, it's nice to
get back to the default single-worktree
directory layout.

(It also helps interop with scripts/tooling
which can't handle the worktree being/moving
leafwards of the repo root.)

This intentionally allows collapsing even if
there are other things in the repo root. The
ideal solution would probably require a flag
to force allowing an unclean collapse, but
but that's not implemented at this time. For
now this is the minimum viable way to cover
the most common need (to return to normal
when down to one cotree and the repo root is
clean) and not force extra work onto people
who want to f.e. keep at least one worktree
while doing the collapse.

P.S. Have I mentioned that I absolutely hate
Bourne shell file globbing? Profound example
of making a design quirk which caters to a
special case and thus has a shape which is
awkward and is a massive pain in the ass for
almost all serious programmatic use.