Add co-tree friendly+safe way to get repository root
    git cotree --show-root

This helps scripts/tools find the repository root,
which with co-trees doubles as the co-tree root
(the directory holding all the co-tree worktrees).
Also might be useful in interactive shells to make
an alias to quickly go back to the root.

The built-in ways are either not obvious, wrong,
or even liable to do the wrong thing in ways
that could be harmful in scripts.

1. `git rev-parse --show-common-dir`/.. could
   point anywhere, depending on the environment
   and possibly other state I haven't considered.
2. `git rev-parse --show-toplevel` will work in
   a worktree (co-tree or otherwise), but it will
   error out outside of the worktree (making it
   useless for finding the repo/co-tree root).
3. Manually searching up to find `.git` has to be
   coded to take obscure edge-case features like
   `GIT_CEILING_DIRECTORIES` into account.

If there's another built-in way that's actually
good for this purpose, I don't know it, and I've
been in the thick of it for a little while now.