Make collapse even more robust against contrived collisions
This handles the unlikely case that the `mktemp`
call inside the worktree generates a name that
already exists in the co-tree/repo root. (This
probably can only happen due to enemy action,
or from a combination of pathologically bad RNG
and `git-cotree` being repeatedly killed or
forced to error out partway through a collapse.)

* Before this commit, if we hit that error,
  we would error out after having already moved
  all the files into the temporary directory,
  leaving the worktree in a garbage state (one
  which is only obviously recoverable if you
  already have decent technical grasp on how
  Git and git-cotree work).

* Now the same error is triggered before any
  of the worktree files are moved into the
  tempdir, because now we lead with the
  tempdir move.

Note: the reason why we `touch` the names
"$temporary" and '.git' is because we need
at least one name in the directory for each
of the `*` and `.*` globs, and those are
the only two known safe-from-collision names
we have.

Otherwise, the globs expand to the literal
'*' and '.*', which is ambiguous with actual
filenames.

`initialize` doesn't have this problem
because "$temporary" and '.git' are in the
worktree at the time that those globs are
expanded for the initialize move loop.

`collapse` now partially has this problem
in its first loop, because "$temporary" is
now moved out of the worktree before the
move loop starts. That's why the `touch
"$temporary"` is moved in front of the 1st
move loop.

`collapse` had and continues to fully have
this problem in the second move loop: it now
benefits from the `case` of the first move
loop letting through "$temporary", but we
still can't move '.git' because `git worktree`
needs that for removing the worktree.