Drop dependency on `git branch --show-current`
One person at the start of this year tried
`git-cotree` on a system that still had a Git
version below 2.22. Luckily it turns out we
can cleanly use older Git plumbing commands,
and the exit codes of that command even let
us distinguish whether HEAD was a branch or
a commit.