Tiny self-explanatory code improvememt
Use a self-descriptive variable name for
the output of the command that resolves
HEAD to the branch it points to (if any).

This matches the name of variable used
for the commit rhat HEAD points to, and
more explicitly says through code what
the intent is for how the co-tree "base"
is chosen: branch if we have one, commit
otherwise.